+++ src/screens/SalesScreen/SalesScreen.tsx
@@ -4,8 +4,8 @@ import { Platform, Text, View } from 'react-native';
 import { StackNavigationOptions } from '@react-navigation/stack';
 import { HeaderButtons, Item } from 'react-navigation-header-buttons';
 
-import HeaderButton from '../components/HeaderButton';
-import { NavigationData } from '../navigation/AppNavigator/SalesNavigator';
+import HeaderButton from '../../components/HeaderButton';
+import { NavigationData } from '../../navigation/AppNavigator/SalesNavigator';
 
 type Props = NavigationData<'SalesList'>;
