+++ src/screens/SalesScreen/index.ts
@@ -1,0 +1,1 @@
+export { default, SalesScreenNavOptions } from './SalesScreen';
